Slight modification to previous commit
toggle-or-{kill,interrupt} now use `eshell-outline--final-prompt-p' to
determine whether to kill or interrupt the process.  This ensures that
the user doesn't actually kill a long-running process while trying to
look at previous output.

Before: if a process is running
After: if a process is running and point is beyond the final prompt